added eslint configuration
