+++ .eslintrc.json
@@ -1,0 +1,125 @@
+{
+    "extends": [
+        "airbnb",
+        "plugin:react/recommended",
+        "prettier",
+        "prettier/react"
+    ],
+    "plugins": ["react"],
+    "env": {
+        "browser": true,
+        "commonjs": true,
+        "es6": true,
+        "node": true,
+        "jest": true
+    },
+    "parser": "babel-eslint",
+    "parserOptions": {
+        "ecmaVersion": 10,
+        "sourceType": "module",
+        "ecmaFeatures": {
+            "jsx": true
+        }
+    },
+    "rules": {
+        "semi": ["error", "always"],
+        "comma-dangle": [
+            "error",
+            {
+                "arrays": "always-multiline",
+                "objects": "always-multiline"
+            }
+        ],
+        "object-curly-spacing": ["error", "always"],
+        "no-multiple-empty-lines": [
+            "error",
+            {
+                "max": 1,
+                "maxEOF": 1
+            }
+        ],
+        "eol-last": ["error", "always"],
+        // "indent": ["error", 4],
+        "no-param-reassign": ["warn"],
+        "arrow-body-style": ["error", "as-needed"],
+        "arrow-parens": [
+            "warn",
+            "as-needed",
+            {
+                "requireForBlockBody": false
+            }
+        ],
+        "arrow-spacing": [
+            "error",
+            {
+                "before": true,
+                "after": true
+            }
+        ],
+        "constructor-super": "error",
+        "no-const-assign": "error",
+        "no-dupe-class-members": "error",
+        "no-duplicate-imports": "error",
+        "no-new-symbol": "error",
+        "no-this-before-super": "error",
+        "no-useless-computed-key": "error",
+        "no-useless-constructor": "warn",
+        "no-useless-rename": [
+            "error",
+            {
+                "ignoreDestructuring": false,
+                "ignoreImport": false,
+                "ignoreExport": false
+            }
+        ],
+        "no-var": "warn",
+        "prefer-arrow-callback": [
+            "warn",
+            {
+                "allowUnboundThis": true
+            }
+        ],
+        "prefer-const": [
+            "error",
+            {
+                "destructuring": "all",
+                "ignoreReadBeforeAssign": true
+            }
+        ],
+        "prefer-template": "error",
+        "require-yield": "error",
+        "rest-spread-spacing": ["error", "never"],
+        "template-curly-spacing": "error",
+        "yield-star-spacing": ["error", "after"],
+        "prefer-spread": "error",
+        "generator-star-spacing": [
+            "error",
+            {
+                "before": false,
+                "after": true
+            }
+        ],
+        "symbol-description": "error",
+        "max-len": ["error", { "code": 100 }],
+        "max-lines-per-function": ["error", 20],
+        "sort-keys": ["error", "asc"],
+
+        "react/no-danger": "error",
+        "react/jsx-pascal-case": "error",
+        "react/jsx-indent": ["error", 4],
+        "react/jsx-indent-props": ["error", 4],
+        "react/jsx-closing-bracket-location": "error",
+        "react/void-dom-elements-no-children": "error",
+        "react/jsx-tag-spacing": "error",
+        "react/jsx-no-literals": "error",
+        "react/jsx-wrap-multilines": "error",
+        "react/jsx-no-comment-textnodes": "error",
+        "jsx-a11y/anchor-is-valid": 0,
+        "import/prefer-default-export": "off"
+    },
+    "settings": {
+        "react": {
+            "version": "latest"
+        }
+    }
+}
