+++ .prettierrc
@@ -1,0 +1,5 @@
+{
+    "trailingComma": "es5",
+    "singleQuote": true,
+    "tabWidth": 4
+}
