--- debug.log
@@ -1,34 +1,0 @@
-[1127/045714.494:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.495:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.496:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.497:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.497:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.497:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.497:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.497:ERROR:crash_report_database_win.cc(469)] failed to stat report
-[1127/045714.497:ERROR:crash_report_database_win.cc(469)] failed to stat report
